Move warnings tests to helper tests
