+++ Lib/test/test_warnings/__init__.py
@@ -5,7 +5,6 @@ from io import StringIO
 import re
 import sys
 import textwrap
-from types import ModuleType, SimpleNamespace
 import unittest
 from test import support
 from test.support import import_helper
@@ -885,146 +884,6 @@ class _WarningsTests(BaseTest, unittest.TestCase):
             with support.swap_item(globals(), '__name__', b'foo'), \
                  support.swap_item(globals(), '__file__', None):
                 self.assertRaises(UserWarning, self.module.warn, 'bar')
-
-    # GH#86298 is part of the migration away from module attributes and toward
-    # __spec__ attributes.  There are several cases to test here.  This will
-    # have to change in Python 3.14 when we actually remove/ignore __loader__
-    # in favor of requiring __spec__.loader.
-
-    @support.cpython_only
-    def test_gh86298_no_loader_and_no_spec(self):
-        bar = ModuleType('bar')
-        del bar.__loader__
-        del bar.__spec__
-        with original_warnings.catch_warnings():
-            # 2022-10-06(warsaw): For backward compatibility with the
-            # implementation in _warnings.c, this can't raise an
-            # AttributeError.  See _bless_my_loader() in _bootstrap_external.py
-            ## self.assertRaises(
-            ##     AttributeError, self.module.warn_explicit,
-            ##     'warning!', RuntimeWarning,
-            ##     'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-            self.assertWarns(
-                RuntimeWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_loader_is_none_and_no_spec(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = None
-        del bar.__spec__
-        with original_warnings.catch_warnings():
-            # 2022-10-06(warsaw): For backward compatibility with the
-            # implementation in _warnings.c, this can't raise an
-            # AttributeError.  See _bless_my_loader() in _bootstrap_external.py
-            ## self.assertRaises(
-            ##     AttributeError, self.module.warn_explicit,
-            ##     'warning!', RuntimeWarning,
-            ##     'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-            self.assertWarns(
-                RuntimeWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_no_loader_and_spec_is_none(self):
-        bar = ModuleType('bar')
-        del bar.__loader__
-        bar.__spec__ = None
-        with original_warnings.catch_warnings():
-            self.assertRaises(
-                ValueError, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_loader_is_none_and_spec_is_none(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = None
-        bar.__spec__ = None
-        with original_warnings.catch_warnings():
-            self.assertRaises(
-                ValueError, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_loader_is_none_and_spec_loader_is_none(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = None
-        bar.__spec__ = SimpleNamespace(loader=None)
-        with original_warnings.catch_warnings():
-            self.assertRaises(
-                ValueError, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_no_spec(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = object()
-        del bar.__spec__
-        with original_warnings.catch_warnings():
-            self.assertWarns(
-                DeprecationWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_spec_is_none(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = object()
-        bar.__spec__ = None
-        with original_warnings.catch_warnings():
-            self.assertWarns(
-                DeprecationWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_no_spec_loader(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = object()
-        bar.__spec__ = SimpleNamespace()
-        with original_warnings.catch_warnings():
-            self.assertWarns(
-                DeprecationWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_loader_and_spec_loader_disagree(self):
-        bar = ModuleType('bar')
-        bar.__loader__ = object()
-        bar.__spec__ = SimpleNamespace(loader=object())
-        with original_warnings.catch_warnings():
-            self.assertWarns(
-                DeprecationWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_no_loader_and_no_spec_loader(self):
-        bar = ModuleType('bar')
-        del bar.__loader__
-        bar.__spec__ = SimpleNamespace()
-        with original_warnings.catch_warnings():
-            self.assertRaises(
-                AttributeError, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
-
-    @support.cpython_only
-    def test_gh86298_no_loader_with_spec_loader_okay(self):
-        bar = ModuleType('bar')
-        del bar.__loader__
-        bar.__spec__ = SimpleNamespace(loader=object())
-        with original_warnings.catch_warnings():
-            self.assertWarns(
-                RuntimeWarning, self.module.warn_explicit,
-                'warning!', RuntimeWarning,
-                'bar.py', 2, module='bar knee', module_globals=bar.__dict__)
 
 
 class WarningsDisplayTests(BaseTest):
